Modularise SAML auth support, making it optional
All the SAML stuff is now in its own sub package and only activated if
auth.saml.enabled is true. This allows the tests to run with the normal
app context, without having to provide everything required by the SAML
auth environment, and keeps the very complex SAML config separate,
avoiding it obfuscating the default security configuration.

Note that the SAML code seems to have come verbatim from:
    https://github.com/thomasdarimont/spring-boot-security-saml-sample/tree/poc/keycloak-saml-idp
It seems to do a lot more than is required to setup a SAML SP; it seems
odd that such a large amount of boilerplate code is required to use
the library. It's not tested, other than that the app context loads and
resolves bean references when SAML is enabled.