Fix ordering issues with modular security config
Adding configurers in the configure(HttpSecurity) method was
problematic as they would end up modifying the built in configurers
which were already run by the time the attempted modifications
happened. The sub configurers now run in-band with the
configure(HttpSecurity) method, so their changes are in place before
the built in configurers are applied.

This also adds support for configuring the WebSecurity and
AuthenticationManagerBuilder builders.

Another issue was that the authentication manager bean would trigger
a dependency loop if a reference to it was needed for a sub configurer
that the top-level WebSecurityConfigurer itself used, as the auth
manager is created by the WebSecurityConfigurer. To fix this, the
AuthenticationManger bean is a proxy which delegates to the real one
created by WebSecurityConfigurer when it runs.