Support repo write isolation & rollback
This implements a very basic method of isolating writes to a git repo
via the GitHelper.writeFilesAndPushOrRollBack() function. Changes are
rolled back if an error occurs, and so long as one GitHelper instance
is used per-repo changes will be isolated as the function is
synchronised.

Nothing to stop an external process (such as another instance of the
app) modifying the repo concurrently though.