WIP: Simplify Item ID handling
The path to an Item from the data root is now used as the Item's ID.

The code was using Item's filename in some situations and the path from
a collection to an Item as an Item's ID. Neither are correct as multiple
Items can have the same filename, and multiple collections can
link to the same Item from different locations. This made it complex to
determine if two collections contained the same item for example.