Include collection ID in toString()
Although I'm not sure that we should have the ID in the model object,
I recon it'd be better to have a generic wrapper object to associate
an ID with the object that represents it. (e.g. the ID cannot really be
added to a model object when deserialising from JSON as it's a higher
level concept than the JSON data the model represents.)