Model content changes as data
This implements a simple data model to represent the desired state of
a model resource:

- Present: With the specified representation
- Absent: Resource should not exist

ModelOps.enforceModelState() takes a list of such states, and applies
them.

Having a data representation for model state changes allows functions
to return a representation of what they want to happen, independant of
actually taking the action. This not only simplifies the implementation
of such functions, but also makes them more straightforward to test.