Return exact type from ModelState.match()
Previously it returned a ModelState<? extends X> wildcard, now it
returns ModelState<X>. The wildcard is accurate, but because instances
are immutable it's safe to use the exact type, as there are no setter
methods that could be passed an incorrect type. X is always the same
or a subtype of the origional type.